Memory/Disk Control
Added a new command-line option, -m, to provide the ability to go back
to the old way of loading everything into memory instead of reading from
a file repeatedly. This gives the user control over the speed vs. memory
usage tradeoff. Also, it allowed for reading from stdin to be put back
in.