Added density proof
Added a text file with a (very rough) proof sketch that for a partition
of the transactions, at least one cell of the partition must have equal
or greater density of a given item set than the density of that item set
in the full transaction set. Provided this proof is correct, this
ensures that the idea behind this branch is a valid one: We can divide
the transaction set up into manageable pieces, run Apriori on each
piece, and then use the union of the patterns identified for each piece
as a superset for the results of the combined run, hopefully thereby
reducing time and memory constraints.