Multiprocessing
Modified the algorithm to be able to subdivide the transactions, run
Apriori on each chunk in a separate process, and then merge the results
back together with a second pass of the Apriori algorithm. Not sure how
well this will work, and it is only the first step in the direction of
making this a distributed algorithm.