Added generational maximum
Added an optional generational_max argument to the pure Python
implementation of the algorithm. This effectively limited the
combinations being considered to a reasonable level. The question
remains, how good are the results produced by this mechanism? It is
possible that important combinations are dropped from the results. A
comparison needs to be made between the results with and without this
generational maximum in place to be certain, but in the meantime, the
algorithm is now usable for large numbers of columns, which is an
improvement.