Correct usage for new savemem.LowMemDict
Changes to the savemem module's interface required minor changes
(simplifications, in fact) to this module. Also, used a built-in dict in
determine_support() and then copied results to item_set_counts, to
minimize disk I/O in the LowMemDict. I suspect that determine_support()
will not generally have high enough memory demands for this to cause an
issue, but that remains to be seen.